build:Add FormatAgnosticOutputStream to cmake build. test:Add test for writing out a reference and a sub-object. refactor:Stream:Set ObjId typedef to be uint32_t rather than size_t for easier streaming. feat:OutputStream,MemoryOutputStream:Add writeUInt32(), writeHeader(), writeFooter(), writeField(). feat:FormatAgnosticOutputStream:Use a BackingStore and a StreamFormat to do the work of writing values to a stream. test:Add test for writing references between objects using MemoryAgnosticOutputStream with a TextFormat and a MemoryBackingStore.

+++ include/io/Stream.h
@@ -15,7 +15,7 @@ namespace dagbase
     {
     public:
         typedef std::uint8_t value_type;
-        typedef std::size_t ObjId;
+        typedef std::uint32_t ObjId;
         typedef void* Ref;
     public:
         virtual ~Stream() = default;
